[Fix] Upgrade Area

+++ Assets/Scripts/PlayerBallCounter.cs
@@ -1,13 +1,14 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Runtime.InteropServices.WindowsRuntime;
+using Unity.VisualScripting;
 using UnityEngine;
 
 public class PlayerBallCounter : MonoBehaviour
 {
     [SerializeField] private BallController _ballController;
 
-    public float stackValue;
+    public int stackValue;
 
     public void BallCountCheck()
     {
@@ -16,7 +17,6 @@ public class PlayerBallCounter : MonoBehaviour
             stackValue += _ballController.balls[i].GetComponent<Ball>().GetValue();
         }
     }
-
     public float GetValue()
     {
         return stackValue;
